+++ src/lib.rs
@@ -93,7 +93,7 @@
 //!
 //! ```toml
 //! [dependencies]
-//! hub75-framebuffer = { version = "0.7.0", features = ["skip-black-pixels"] }
+//! hub75-framebuffer = { version = "0.8.0", features = ["skip-black-pixels"] }
 //! ```
 //!
 //! ### `esp32-ordering` Feature (required for original ESP32 only)
@@ -105,7 +105,7 @@
 //!
 //! ```toml
 //! [dependencies]
-//! hub75-framebuffer = { version = "0.7.0", features = ["esp32-ordering"] }
+//! hub75-framebuffer = { version = "0.8.0", features = ["esp32-ordering"] }
 //! ```
 //!
 //! ### `defmt` Feature