+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - ReleaseDate
 
+## [0.8.0] - 2026-05-01
+
 ### ⚠️ Breaking
 
 * removed unneeded const generics from the FrameBuffer, MutableFrameBuffer, and FrameBufferOperations traits
@@ -79,7 +81,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * initial version
 
 <!-- next-url -->
-[Unreleased]: https://github.com/liebman/hub75-framebuffer/compare/v0.7.0...HEAD
+[Unreleased]: https://github.com/liebman/hub75-framebuffer/compare/v0.8.0...HEAD
+[0.8.0]: https://github.com/liebman/hub75-framebuffer/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/liebman/hub75-framebuffer/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/liebman/hub75-framebuffer/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/liebman/hub75-framebuffer/compare/v0.4.2...v0.5.0